backend/requirements.txt
# Core Backend
sqlalchemy<2.0
fastapi>=0.100.0
pydantic>=2.3.0
uvicorn>=0.15.0,<0.16.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6
openai>=1.2.0
alembic>=1.6.5,<1.7.0
psycopg2-binary>=2.9.1,<3.0.0
python-dotenv==1.0.0
pydantic-settings<=3.0.0, >=2.4.0
email-validator>=2.0.0

# Async, Testing, Web
pytest>=6.2.4,<6.3.0
pytest-asyncio>=0.15.1,<0.16.0
httpx>=0.23.0,<1.0.0
websockets>=10.0,<11.0
redis==5.0.1

# LangChain & Agents
langchain>=0.3.26
langchain-community>=0.3.26
langchain-openai>=0.3.24
langgraph>=0.4.8
unstructured>=0.11.0
tavily-python>=0.3.0
python-magic>=0.4.27
pyautogen>=0.2.0
e2b-code-interpreter>=0.0.10

# OCR and Document Processing
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
opencv-python>=4.8.0
amazon-textract-caller>=0.2.0
pypdf>=3.17.0
python-docx>=1.1.0

# AWS
boto3>=1.35.0
botocore>=1.35.0

# Web Search Tools
google-search-results>=2.4.2
serpapi>=0.1.5

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# ML/AI Libraries
torch>=2.0.0
transformers>=4.33.0
sentence-transformers>=2.2.2

# OpenTelemetry for Logging
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-sqlalchemy>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-redis>=0.41b0
opentelemetry-exporter-otlp>=1.20.0

# Guardrails AI
guardrails-ai>=0.6.6

# Pin setuptools to avoid pkg_resources deprecation warning from guardrails
setuptools<81 